small changes to DMHardware, Did some tests with WobbleGoalPickUp and it works well. If you run WobbleGoalPickUp it will show up as "ArmTest."
