+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DMHardware.java
@@ -45,7 +45,7 @@ public class DMHardware {
         intakeServoRight = hwMap.servo.get("intake_servo_right");
         backLeft.setDirection(DcMotor.Direction.REVERSE);
         frontLeft.setDirection(DcMotor.Direction.REVERSE);
-        wobbleGoalClaw.setPosition(0.5);
+        wobbleGoalClaw.setPosition(0);
        webcamName = hwMap.get(WebcamName.class, "Webcam 1");
     }
